Close blue/red done 02/18/24
